# docs/swagger.yaml
definitions:
...
  permission.Permission:
    properties:
      category:
        type: string
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  role.CreateRequest:
    properties:
      name:
        example: Администратор
        type: string
      permission_ids:
        example:
        - 1
        - 2
        - 3
        - 4
        - 5
        items:
          type: integer
        type: array
    required:
    - name
    type: object
  role.ErrorResponse:
    properties:
      error:
        example: Описание ошибки
        type: string
    type: object
  role.Role:
    properties:
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/permission.Permission'
        type: array
    type: object
  role.SuccessResponse:
    properties:
      message:
        example: Операция выполнена успешно
        type: string
    type: object
  role.UpdateRequest:
    properties:
      name:
        example: Модератор
        type: string
      permission_ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      role_id:
        example: 1
        type: integer
    required:
    - name
    type: object
  user.CreateRequest:
    properties:
      full_name:
        example: Admin User
        type: string
      password:
        example: "123456"
        type: string
      role_id:
        example: 1
        type: integer
      username:
        example: admin
        type: string
    type: object
  user.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        $ref: '#/definitions/role.Role'
      roleID:
        format: int64
        type: integer
      username:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: Возвращает список всех ролей с их разрешениями
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/role.Role'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/role.ErrorResponse'
      summary: Получить список всех ролей
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Создает новую роль с указанными разрешениями
      parameters:
      - description: Данные для создания роли
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/role.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/role.ErrorResponse'
      summary: Создать новую роль
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет роль по указанному ID
      parameters:
      - description: ID роли
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/role.ErrorResponse'
      summary: Удалить роль
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Возвращает роль по указанному ID со списком разрешений
      parameters:
      - description: ID роли
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/role.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/role.ErrorResponse'
      summary: Получить роль по ID
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Обновляет информацию о роли и её разрешения
      parameters:
      - description: ID роли
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления роли
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/role.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/role.ErrorResponse'
      summary: Обновить роль
      tags:
      - roles
  /roles/{id}/permissions:
    get:
      consumes:
      - application/json
      description: Возвращает список разрешений для указанной роли
      parameters:
      - description: ID роли
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/permission.Permission'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/role.ErrorResponse'
      summary: Получить разрешения роли
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Полностью заменяет список разрешений для указанной роли
      parameters:
      - description: ID роли
        in: path
        name: id
        required: true
        type: integer
      - description: Список ID разрешений
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/role.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/role.ErrorResponse'
      summary: Обновить разрешения роли
      tags:
      - roles
  /users:
    get:
      description: Возвращает список всех пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "500":
          description: internal error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создает нового пользователя
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: bad request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      description: Удаляет пользователя по ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить пользователя
      tags:
      - users
    get:
      description: Возвращает пользователя по ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет данные пользователя
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: bad request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить пользователя
      tags:
      - users
swagger: "2.0"